modified to get two topic models (male, female); + MAJOR updates
the code now trains 2 models, 1 for the male corpus and 1 for the female corpus. The dataset is the most up-to-date, but NOT the final one. The retrieving of characters' names has been partially automated and a function for creating a csv file with the most important results has been added