deleted saved texts preprocessed with old criteria
Since preprocessing texts takes much time, a specific cell saves the output of the preprocessing phase in .txt files saved in m_preprocess and f_preprocess, in order to later directly use these by de-commenting a specific cell and skipping the preprocessing cell. The texts stored had been preprocessed with a previous version of the corpus, so i removed them.