new preprocessing methods + new modeling strategy
In tm_process.ipynb I model the WHOLE corpus, processing it by only including nouns that are not included in the stopwords list (currently the stopwords do not include gendered words as 'man' or 'wife'). Great results! We just need to adjust the num_topics parameter by trial and error! So far n=20 and n=40 have been tested